agent/pyproject.toml
[project]
name = "haremovie_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.13.0",
    "google-genai>=1.32.0",
    "loguru>=0.7.3",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
]
